Revert "Login service"
