--- 2-add-ad-b2c/forms/TheReviewer.Core/Services/IIdentityService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.Identity.Client;
-
-namespace TheReviewer.Core
-{
-    public interface IIdentityService
-    {
-        Task<AuthenticationResult> Login();
-        Task<AuthenticationResult> GetCachedSignInToken();
-        UIParent UIParent { get; set; }
-    }
-}
