+++ 2-add-ad-b2c/forms/TheReviewer.Core/TheReviewer.Core.csproj
@@ -7,14 +7,13 @@
   <ItemGroup>
     <PackageReference Include="Refractored.MvvmHelpers" Version="1.3.0" />
     <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+    <PackageReference Include="Microsoft.Identity.Client" Version="1.1.0-preview" />
     <PackageReference Include="Xamarin.Forms" Version="2.5.0.121934" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="1.1.1-alpha0417" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Pages\" />
     <Folder Include="ViewModels\" />
     <Folder Include="Models\" />
     <Folder Include="Services\" />
-    <Folder Include="Helpers\" />
   </ItemGroup>
 </Project>
