+++ 2-add-ad-b2c/forms/iOS/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.CSharp" version="4.4.1" targetFramework="xamarinios10" />
-  <package id="Microsoft.Identity.Client" version="1.1.1-alpha0417" targetFramework="xamarinios10" />
+  <package id="Microsoft.CSharp" version="4.4.0" targetFramework="xamarinios10" />
+  <package id="Microsoft.Identity.Client" version="1.1.0-preview" targetFramework="xamarinios10" />
   <package id="Microsoft.NETCore.Platforms" version="2.0.1" targetFramework="xamarinios10" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="xamarinios10" />
   <package id="NETStandard.Library" version="2.0.1" targetFramework="xamarinios10" />
